Checkpoint.
